requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
click==8.1.6
colorama==0.4.6
dnspython==2.4.2
fastapi==0.101.0
fastapi-csrf-protect==0.3.2
gunicorn==21.2.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
motor==3.2.0
packaging==23.1
passlib==1.7.4
pydantic==2.1.1
pydantic_core==2.4.0
PyJWT==2.8.0
pymongo==4.4.1
python-decouple==3.8
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2